% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSeqLib.R
\name{addReads}
\alias{addReads}
\title{addReads}
\usage{
addReads(.Object, reads, qual, qname)
}
\arguments{
\item{reads}{character vector of [ACGT]* OR data.table with columns $seq and optional fields $qual, $qname}

\item{qual}{optional character vector of ASCII quality strings, nchar(qual) and nchar(reads) must be identical (default "~")}

\item{qname}{character vector of qname (default reads are numbered 1:length(reads))}
}
\description{
Adds reads to Fermi object, clearing contigs field.
}
\author{
Marcin Imielinski
}
