% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSeqLib.R
\name{BWA}
\alias{BWA}
\title{BWA}
\usage{
BWA(fasta = NULL, seq = NULL, seqname = "myseq", mc.cores = 1,
  hardclip = FALSE, keep_sec_with_frac_of_primary_score = 0.9,
  max_secondary = 10)
}
\description{
BWA
}

