% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSeqLib.R
\docType{methods}
\name{$<-}
\alias{$<-}
\alias{$<-,BWA-method}
\title{$<-}
\usage{
\S4method{$}{BWA}(x, name) <- value
}
\arguments{
\item{x}{\code{BWA} object to alter \code{params} field of}

\item{name}{\code{params} field to alter}

\item{value}{New value}
}
\description{
Setting query params of BWA object 

Usage:
bwa$mc.cores = 10
bwa$max_secondary = 5
}
\author{
Marcin Imielinski
}
