% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSeqLib.R
\name{fermi_initialize}
\alias{fermi_initialize}
\title{fermi_initialize}
\usage{
\S4method{initialize}{Fermi}(
  .Object,
  reads = NULL,
  qual = NULL,
  qname = NULL,
  assemble = FALSE,
  correct = TRUE
)
}
\arguments{
\item{reads}{character vector of [ACGT] OR data.table with columns $seq and optional fields $qual, $qname}

\item{qual}{optional character vector of ASCII quality strings, nchar(qual) and nchar(reads) must be identical (default "~")}

\item{qname}{character vector of qname (default reads are numbered 1:length(reads))}

\item{assemble}{logical flag of whether to immediately assemble}

\item{correct}{logical flag of whether to error correct if we immediately assemble}
}
\description{
initialize fermi
}
\author{
Marcin Imielinski
}
