% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSeqLib.R
\name{load_index}
\alias{load_index}
\title{load_index}
\usage{
\S4method{initialize}{BWA}(
  .Object,
  fasta = NULL,
  seq = NULL,
  seqname = NULL,
  mc.cores = 1,
  hardclip = FALSE,
  keep_sec_with_frac_of_primary_score = 0.9,
  max_secondary = 10
)
}
\description{
load_index
}
\author{
Marcin Imielinski
}
