% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSeqLib.R
\name{query}
\alias{query}
\title{query}
\usage{
query(.Object, ...)
}
\arguments{
\item{qstring}{character vector of DNA reads, with optional names specifying qnames}

\item{qname}{optional vector of read names to provide with query}

\item{hardclip}{logical flag whether to hardclip alignments (default FALSE)}

\item{keep_sec_with_frac_of_primary_score}{BWA threshold for keeping secondary alignments}

\item{max_secondary}{integer max number of alignments to keep}
}
\description{
query

Query character vector of DNA reads using BWA.
}
\author{
Marcin Imielinski
}
